renamed Trigger classes
