+++ core/src/test/java/apoc/trigger/TriggerRestartTest.java
@@ -39,7 +39,7 @@ public class TriggerRestartTest {
         databaseManagementService = new TestDatabaseManagementServiceBuilder(store_dir.getRoot().toPath()).build();
         db = databaseManagementService.database(GraphDatabaseSettings.DEFAULT_DATABASE_NAME);
         ApocConfig.apocConfig().setProperty("apoc.trigger.enabled", "true");
-        TestUtil.registerProcedure(db, Trigger.class, TriggerDeprecatedProcedures.class);
+        TestUtil.registerProcedure(db, TriggerNewProcedures.class, Trigger.class);
     }
 
     @After
